mcp_server/requirements.txt
# Stockly MCP Server Requirements
# متطلبات خادم MCP للمحاسب الذكي

# MCP Server
fastmcp>=0.1.0

# HTTP Requests
requests>=2.32.0

# Data Validation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Logging (اختياري)
colorlog>=6.7.0

# Development Dependencies (اختياري)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
